MEX-1156 MultiForward Recall: Failed to Suspend Transaction error occurs for item with max quantity & price.
git-svn-id: http://svnsin.ncr.com/svn/repos/pa/res/RES%20MeX/Base/resTransaction/branches/3.3.2@79374 9234da4d-b7fe-4b44-acf3-10f21a033f33

+++ src/ncr/res/mobilepos/journalization/dao/SQLServerPosLogDAO.java
@@ -2184,7 +2184,7 @@ public class SQLServerPosLogDAO extends AbstractDao implements IPosLogDAO {
                     ("false".equals(transaction.getTrainingModeFlag())) ? 0 : 1);
             saveForwardPosLogPrepStmnt.setString(SQLStatement.PARAM8, transaction.getBeginDateTime());
             saveForwardPosLogPrepStmnt.setString(SQLStatement.PARAM9, transaction.getOperatorID().getValue());
-            saveForwardPosLogPrepStmnt.setInt(SQLStatement.PARAM10, Integer.parseInt(total));
+            saveForwardPosLogPrepStmnt.setLong(SQLStatement.PARAM10, Long.parseLong(total));
             saveForwardPosLogPrepStmnt.setString(SQLStatement.PARAM11, transaction.getWorkStationID().getValue());
             saveForwardPosLogPrepStmnt.setString(SQLStatement.PARAM12, "0");
             saveForwardPosLogPrepStmnt.setString(SQLStatement.PARAM13, posLogXml);
